adjustment: separated type declaration for segmented buttons
