+++ src/components/SegmentedTabs/type.ts
@@ -1,0 +1,4 @@
+export interface SegmentButtons {
+  tabName: string;
+  component?: React.ReactNode;
+}
