+++ src/pages/notifications/components/NotifComponent.tsx
@@ -1,5 +1,5 @@
 import { Block, Icon, Link } from 'framework7-react';
-import { NotificationDataType } from '../utils';
+import { NotificationDataType } from '../type';
 
 export const NotifComponent = ({ data }: { data: NotificationDataType[] }) => {
   return (
