+++ src/pages/notifications/type.ts
@@ -1,0 +1,8 @@
+export interface NotificationDataType {
+  id: number;
+  title: string;
+  type: string;
+  content: string;
+  date?: string;
+  time?: string;
+}
